Fix: Extract abstract test case to avoid duplication
